# Reduce startup latency by re-using OkHttp client