Before this patch:
```
09-06 00:53:21.707  18649    18670          JSONConnector  D  internalLogin: 1071ms
09-06 00:53:22.647  18649    18670          JSONConnector  D  getHeadlines: 2011ms
09-06 00:53:23.664  18649    18670          JSONConnector  D  getHeadlines: 1012ms
09-06 00:53:24.688  18649    18670          JSONConnector  D  getCategories: 984ms
09-06 00:53:25.768  18649    18670          JSONConnector  D  getFeeds: 1072ms
```
After this patch:
```
09-06 00:52:58.534  18415    18436          JSONConnector  D  internalLogin: 1135ms
09-06 00:52:58.854  18415    18436          JSONConnector  D  getHeadlines: 1455ms
09-06 00:52:59.149  18415    18436          JSONConnector  D  getHeadlines: 291ms
09-06 00:52:59.458  18415    18436          JSONConnector  D  getCategories: 274ms
09-06 00:52:59.805  18415    18436          JSONConnector  D  getFeeds: 338ms
```
43% reduction in terms of latency :smiley: 

Ping latency for my server: ~160ms. I use TLS 1.3 + HTTP/2 on Apache.