@yan12125 2020-09-08
Hmm, seems after this patch changing proxy and lazy server settings will not work correctly until restarting the App. Is that acceptable for you?