@nilsbraden 2020-09-09
Thank you for the patch!

> Hmm, seems after this patch changing proxy and lazy server settings will not work correctly until restarting the App. Is that acceptable for you?

Hm this would raise problems also when using the Wifi-Preferences since in that case proxy settings can change with every new Wifi, at least theoretically. I'll try to add a check in JSONConnector to reload OkHttp when any of these settings changed.